requirements.txt
# Core Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8 for object detection

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Data Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# ML Utilities
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Configuration & Logging
PyYAML>=6.0
python-box>=7.0.0  # Easy access to dict values
ensure>=1.0.2
pydantic>=2.0.0

# Web Framework (for deployment)
Flask>=2.3.0
Flask-Cors>=4.0.0
gunicorn>=21.0.0

# API & Serialization
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Utilities
tqdm>=4.65.0  # Progress bars
python-dotenv>=1.0.0
joblib>=1.3.0

# Monitoring & Logging
tensorboard>=2.13.0
wandb>=0.15.0  # Optional: for experiment tracking

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type Checking
mypy>=1.4.0
types-PyYAML>=6.0.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Deployment
docker>=6.1.0


# Additional utilities
requests>=2.31.0
pathlib>=1.0.1
